float to double in importance weights change calculation of entropy gradient to reduce numeric errors
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/branches/josiah@2273 1f5c12ca-751b-0410-a591-d2e778427230

+++ josiah/Gibbler.cpp
@@ -421,14 +421,14 @@ void Sampler::Run(Hypothesis* starting, const TranslationOptionCollection* optio
         m_operators[j]->doIteration(sample,*options);
       }
       //importance weight
-      float totalImpWeight = 0;
+      double totalImpWeight = 0;
       ScoreComponentCollection deltaFV;
       for (Josiah::feature_vector::const_iterator j = sample.extra_features().begin(); j != sample.extra_features().end(); ++j) {
-        float score = (*j)->getImportanceWeight(sample);
+        double score = (*j)->getImportanceWeight(sample);
         const ScoreProducer& sp = (*j)->getScoreProducer();
         const StaticData& staticData = StaticData::Instance();
         const ScoreIndexManager& sim = staticData.GetScoreIndexManager();
-        float scoreWeight = StaticData::Instance().GetAllWeights()[sim.GetBeginIndex(sp.GetScoreBookkeepingID())];
+        double scoreWeight = StaticData::Instance().GetAllWeights()[sim.GetBeginIndex(sp.GetScoreBookkeepingID())];
         VERBOSE(2, "Score producer: " << sp.GetScoreProducerDescription() << " Weight: " << scoreWeight << endl)
         totalImpWeight += scoreWeight*score;
         //set the weight in the sample  
@@ -461,7 +461,7 @@ void PrintSampleCollector::collect(Sample& sample)  {
 
 
 
-void SampleCollector::addSample( Sample & sample, float importanceWeight ) {
+void SampleCollector::addSample( Sample & sample, double importanceWeight ) {
   if (m_importanceWeights.empty()) {
     m_totalImportanceWeight = importanceWeight;
   } else {
